// src/types/beats/starknet.ts
// Auto-generated by scripts/generate-block-types.ts
// Generated: 2026-03-17
// Sources:
//   https://jiti.indexing.co/networks/starknet/7797628
//   https://jiti.indexing.co/networks/starknet/7797528
//   https://jiti.indexing.co/networks/starknet/7796628
// Samples: 3
export interface StarknetBlockL1GasPrice {
  price_in_fri: string;
  price_in_wei: string;
}

export interface StarknetBlockL1DataGasPrice {
  price_in_fri: string;
  price_in_wei: string;
}

export interface StarknetBlockL2GasPrice {
  price_in_fri: string;
  price_in_wei: string;
}

export interface StarknetBlockTransactionsResourceBoundsL1Gas {
  max_amount: string;
  max_price_per_unit: string;
}

export interface StarknetBlockTransactionsResourceBoundsL2Gas {
  max_amount: string;
  max_price_per_unit: string;
}

export interface StarknetBlockTransactionsResourceBoundsL1DataGas {
  max_amount: string;
  max_price_per_unit: string;
}

export interface StarknetBlockTransactionsResourceBounds {
  l1_gas: StarknetBlockTransactionsResourceBoundsL1Gas;
  l2_gas: StarknetBlockTransactionsResourceBoundsL2Gas;
  l1_data_gas: StarknetBlockTransactionsResourceBoundsL1DataGas;
}

export interface StarknetBlockTransactionsReceiptActualFee {
  amount: string;
  unit: string;
}

export interface StarknetBlockTransactionsReceiptEvents {
  from_address: string;
  keys: string[];
  data: string[];
}

export interface StarknetBlockTransactionsReceiptExecutionResources {
  l1_gas: number;
  l2_gas: number;
  l1_data_gas: number;
}

export interface StarknetBlockTransactionsReceipt {
  type: string;
  transaction_hash: string;
  actual_fee: StarknetBlockTransactionsReceiptActualFee;
  execution_status: string;
  finality_status: string;
  block_hash: string;
  block_number: number;
  messages_sent: unknown[];
  events: StarknetBlockTransactionsReceiptEvents[];
  execution_resources: StarknetBlockTransactionsReceiptExecutionResources;
  revert_reason?: string;
}

export interface StarknetBlockTransactions {
  transaction_hash: string;
  type: string;
  version: string;
  nonce: string;
  sender_address: string;
  signature: string[];
  calldata: string[];
  resource_bounds: StarknetBlockTransactionsResourceBounds;
  tip: string;
  paymaster_data: unknown[];
  account_deployment_data: unknown[];
  nonce_data_availability_mode: string;
  fee_data_availability_mode: string;
  receipt: StarknetBlockTransactionsReceipt;
}

export interface StarknetBlock {
  status: string;
  block_hash: string;
  parent_hash: string;
  block_number: number;
  new_root: string;
  timestamp: number;
  sequencer_address: string;
  l1_gas_price: StarknetBlockL1GasPrice;
  l1_data_gas_price: StarknetBlockL1DataGasPrice;
  l1_da_mode: string;
  starknet_version: string;
  l2_gas_price: StarknetBlockL2GasPrice;
  transactions: StarknetBlockTransactions[];
  _network: string;
  _vm: string;
}